+++ src/main/java/com/braintreegateway/PaymentMethodNonceDetails.java
@@ -4,15 +4,25 @@ import com.braintreegateway.util.NodeWrapper;
 
 public class PaymentMethodNonceDetails {
     private String cardType;
-    private String lastTwo;
+    private String cardholderName;
+    private String correlationId;
+    private String dpanLastTwo;
+    private String email;
     private String lastFour;
+    private String lastTwo;
+    private String paymentInstrumentName;
     private String username;
     private String venmoUserId;
 
     public PaymentMethodNonceDetails(NodeWrapper node) {
         cardType = node.findString("card-type");
-        lastTwo = node.findString("last-two");
+        cardholderName = node.findString("cardholder-name");
+        correlationId = node.findString("correlation-id");
+        dpanLastTwo = node.findString("dpan-last-two");
+        email = node.findString("email");
         lastFour = node.findString("last-four");
+        lastTwo = node.findString("last-two");
+        paymentInstrumentName = node.findString("payment-instrument-name");
         username = node.findString("username");
         venmoUserId = node.findString("venmo-user-id");
     }
@@ -21,12 +31,32 @@ public class PaymentMethodNonceDetails {
         return cardType;
     }
 
+    public String getCardholderName() {
+        return cardholderName;
+    }
+
+    public String getCorrelationId() {
+        return correlationId;
+    }
+
+    public String getDpanLastTwo() {
+        return dpanLastTwo;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
     public String getLastTwo() {
         return lastTwo;
     }
 
     public String getLastFour() {
         return lastFour;
+    }
+
+    public String getPaymentInstrumentName() {
+        return paymentInstrumentName;
     }
 
     public String getUsername() {