+++ CHANGELOG.md
@@ -1,6 +1,8 @@
+## unreleased
+* Add `GrantedPaymentMethodRevoked` webhook response objects
+
 ## 2.86.0
 * Add `authorizationExpiresAt` to `Transaction`
-* Add `GrantedPaymentMethodRevoked` webhook response objects
 
 ## 2.85.0
 * Add venmo account details to payment method nonce model